requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Limiter==3.12
Flask-Mail==0.10.0
flask-mongoengine==1.0.0
Flask-WTF==1.2.2
google_search_results==2.4.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mongoengine==0.29.1
ordered-set==4.1.0
packaging==25.0
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.12.1
python-dotenv==1.1.0
requests==2.32.3
rich==13.9.4
serpapi==0.1.5
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
